requirements.txt
discord.py>=2.3.2
chat-exporter>=2.7.0
python-dotenv>=1.0.0
psutil>=5.9.0
aiohttp>=3.9.0
asyncio>=3.4.3
python-dateutil>=2.8.2
typing-extensions>=4.8.0